user_auth/requirements.txt
aiohttp==3.6.2
aiodns==2.0.0
cchardet==2.1.6
aiohttp-swagger==1.0.15
aiohttp-devtools==0.13.1
alembic==1.4.2
SQLAlchemy==1.3.18
psycopg2-binary==2.8.5
aioredis==1.3.1
aiopg==1.0.0
argon2-cffi==20.1.0
PyJWT==1.7.1
aiohttp-jwt==0.6.1
